+++ src/main/java/com/spicelang/intellij/spice/debug/SpiceDebugConfigurationFactory.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2022-2026 ChilliBits. All rights reserved.
+ */
+
+package com.spicelang.intellij.spice.debug;
+
+import com.intellij.execution.configurations.ConfigurationFactory;
+import com.intellij.execution.configurations.ConfigurationType;
+import com.intellij.execution.configurations.RunConfiguration;
+import com.intellij.openapi.components.BaseState;
+import com.intellij.openapi.project.Project;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class SpiceDebugConfigurationFactory extends ConfigurationFactory {
+
+    protected SpiceDebugConfigurationFactory(ConfigurationType type) {
+        super(type);
+    }
+
+    @Override
+    public @NotNull String getId() {
+        return SpiceDebugRunConfigurationType.ID;
+    }
+
+    @NotNull
+    @Override
+    public RunConfiguration createTemplateConfiguration(@NotNull Project project) {
+        return new SpiceDebugRunConfiguration(project, this, "Spice (Native)");
+    }
+
+    @Nullable
+    @Override
+    public Class<? extends BaseState> getOptionsClass() {
+        return SpiceDebugRunConfigurationOptions.class;
+    }
+}